+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-## Cockpit Ceph-Deploy 0.1.5-19
+## Cockpit Ceph-Deploy 0.2.0-1
 
-* adding warnadded warning indicators if parameters are changeimportant parameters are changed
+* added primitive warning messages when parameters used to deploy are modified